refactor: group models by role instead of by sender
